fixed typescript compatibility error with route.ts name in the api/auth/[...nextauth] folder
