+++ src/utils/serverUtils.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import authOptions from "@/app/api/auth/[...nextauth]/options";
 import { getServerSession } from "next-auth/next";
 import { AuthSession } from "@/types/types";
